requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
websockets==12.0
requests==2.32.3
